Add get_bc_epochs and option to specify BC filters
The bolometric corrections are different for different pairs of filters,
so I needed a way of fetching epochs with the requirement that both of a
specified pair of filters was used on that day.

get_bc_epochs does that, creating an array to hold the valid epochs.
This will get used by lbol_bc_bh09 to generate Lbol on those epochs.