Fix non-monotonic JD bug in lbol_bc_bh09
The BC lightcurves generated incorrect results if used with a dataset in
the HDF5 file that was non-monotonic (as was, for example, the dataset
from SN 2000cb, which combined two separate sets of published
observations.)

The problem was the way the data was being passed back-and-forth in the
code. A list of sorted dates was stepped through, but other arrays which
held the color and V magnitude were out of order, so when referenced,
data from the wrong days was being combined, and the lightcurves were
incorrect.

This problem has now been fixed by modifying the get_ functions for
color and V mangitude to take the JD as an argument, ensuring that the
correct values are used in the calculation of Lbol