Add blackbody synthetic photometry to HDF5 file
I created a grid of BB models with the same temperatures and
luminosities as the WD models, set them at a distance of 10pc, and then
used Eddie's synthetic photometry code to calculate magnitudes.

This data will be used to validate the Lbol calculated from
quasi-bolometric and direct integration methods.