Fix plaintext output
The direct integration method was outputting the temperature and angular
radius as well as the information I actually want in the plaintext
output. This was done for debugging purposes, but isn't necessary
anymore.

I also didn't have the columns of the output labeled properly, so I've
corrected that and added units to the output (just to make things
clear.)