Move h5file from global to local variable
I had defined the h5file to be a global variable in the script
hdf5_io.py. This makes the code hard to test, as simply importing the
script will cause a new hdf5 file to appear out of nowhere.

Now, the two main functions in the scipt take a variable for the name of
the file, allowing for more flexible testing to occur on an example file
that I will keep in the tests/ directory.