Add tests to check WD temperature fits in 2.7 & 3
I'm slowly ruling out the possible sources of the direct integration
luminosity difference between Python 2.7 and Python 3.
One possibility was that the scipy routines that fit a blackbody to the
models were working differently.

The temperatures agree to 4 decimal places.
Not perfect, but I need to do more work to see if this is the source of
the problem.