Fix hidden integer division in planck integral
The last term in the infinite series approximation of the Planck
integral is 6/n^4. This was coded as 6/n**4, but since both 6 and n are
integers, Python 2.7 was doing integer division. Python 3 defaults to
floating-point division, so this was the source of the discrepancy
between the results obtained from the code when running under the two
different interpreters.