Added photometry "zipping" functionality
For lack of a better term, I'm using the term "zip" to describe the
process of merging together different photometric datasets, which can
use different filter and which can have been taken at different times.

The process of zipping sorts the full dataset by JD, then combines data
from the same 'night' (where night has been defined to be a cluster of
observations within 0.3 days of one another) and assigns those
observations the same JD (the average of the JDs of the observations.)

For cases when there were multiple detections using the same filter on
the same night, those magnitudes are averaged together and the errors
added in quadrature.