Clean slate
Wiping out the previous code, not because I've lost my mind, but because
I want to learn. I'm going to approach the problem that SuperBoL
attempts to solve using test-driven development and object orientation.
I'm not doing this because those are the only correct ways to solve a
problem, but because I want to understand how those methods work and how
to use them to solve *this* problem.

So, with any luck, this is going to be fun!