Observation calculates uncertainty in flux
Added a method to the Observation class which calculates the uncertainty of
the flux as calculated by the convert_to_flux() method. This needs to be
refactored to clean up the class, and I think the whole flux conversion
can be factored out into its own class, but I'm taking baby steps.