Refactor MagnitudeToFluxConverter
Instead of passing the observation to the converter when it is created,
the convert() method takes an observation from the outside.

The methods for calculating the flux and for calculating the flux
uncertainty are now more straightforward. They now take all their
arguments directly, so testing is easier.

I also removed some duplicated tests from the test suite. Since the
converter object does all the heavy lifting, the bulk of the unit tests
now lie there. The observed magnitude tests are limited to checking that
the whole module is working as intended.