Process fluxes before trapezoidal integration
Rather than try to deal with duplicated observation in the trapezoidal
integration routine itself, I factored that functionality out to
independent functions. Those functions take the list of flux objects,
iterate through it to find the fluxes at each wavelength, and combine
repeated fluxes using a simple averaging.

In the future, it might be better to use a weighted mean method to
combine the fluxes, but this will be a simple change now that the
functinality is separate from the integration routine.