Implented linear interpolation for making SEDs
If an SED is missing an observation, this interpolation module will
estimate the missing flux as long as there are observed fluxes at the
same wavelength before and after the missing flux. If the "gap" between
observed fluxes around the missing flux is less than or equal to 2 days,
the flux is estimated using linear interpolation. If it is greater than
2 days, it is not filled in.

TODO: In the future, it might be nice to have multiple interpolation
schemes available that can be used via dependency injection or (in the
case of of Lyman et. al. (2014)) if there are gaps of more than 2 days,
the flux could be estimated using cubic splines.