Add planck and fit_blackbody modules
I'm copying these modules and their tests from the earlier version of
the code. The only significant change I had to make in order for the
tests to run was to change the number of decimal places to which the
assertAlmostEqual() functions were comparing. There is probably a more
elegant solution for this, but that's for another time.