Modify yield function to work with magnitudes
This revealed an interesting bug in the way the yield function was
working. Because I was forming a set(), which is unordered, all the unit
testing that I had been doing were just getting lucky that the ordering
of the set was the same every time. For whatever reason, that wasn't
true when I was forming a set of the band names. I've fixed this by
replacing the set with a sorted(list(set())). Probably a cleverer way to
do this, but at least the fix is isolated, and I was swap it out with
something better in the future.