Add get_observed_band_names
This is the first step to interpolating for missing observations. The
only difference between this code and the SED code that preceded it is
that it returns the band names sorted alphabetically. This is to make
the unit tests pass, but it means that I cannot assume that the
observations are in wavelength order when I'm processing them further.