ObservedMagnitudes take Bands as arguments
I had written the test code and the photometry module forgetting that
the ObservedMagnitude takes a Band object as an argument, not just a
string with the name of the band. I've updated the test code and the
photometry module to fix this.

This also necessitated adding an __eq__(): method to the Band object, in
order to be able to test for the equality of two observed magnitudes.