Add function to calculate BC luminosity
This is the main interface for calculating a BC luminosity.

In the future, I should get rid of some of this duplication between the
calculate bc luminosity code and the other luminosity calculation. This
would probably involve generalizing the methods that are currently
implemented in the sed and photometry modules into methods that worked
on 'observations' be they fluxes or magnitudes.