Modify lqbol.py to expect an SED as input instead of creating one
Originally, get_quasi_bolometric_flux() took the raw list of
monochromatic fluxes as its input, then as a first step converted
those to an SED. This would take care of cases where there were
multiple observations through the same filter on the same date.

This caused an unnecessary dependence, and was making unit testing
harder, so I've modified get_quasi_bolometric_flux() to expect an
SED instead of creating one. Now I need to make sure that an SED
is always created from the observations before the call to lqbol,
and probably add some checks to get_quasi_bolometric_flux() to
make sure that the list of fluxes being input is an SED.