Starting to fix the bugs in SED construction and flux averaging
Started by identifying the JD that the SN 2000cb data is throwing
and error. Looks like on JD 51683.25 the V and R data isn't being
averaged properly.

Wavelength list:  [4380.0, 5450.0, 6410.0, 6410.0, 7980.0, 7980.0]
Flux list:  [7.393729782818236e-16, 9.011821574832454e-16,
             7.842368381009181e-16, array(7.83623551e-16),
	     4.737285387956838e-16, array(4.7342445e-16)]

This should only be a 4-observation SED (U, B, V, and R).
Instead, the V and R are showing up twice in the wavelength and
flux lists, once and then again as this weird array() thing.

Next step is to re-create the error in a regression test and
reproduce it intentionally.