Fixed the missing flux interpolation problem. Sorta.
So the problem was that fluxes that happened on the same day were
being grouped together, but they kept their original times.

When the interpolation step happened, the assumption was that
the fluxes in each SED shared the same time, but because they
kept their original times, they were in the SED but their times
still varied (observations with times 1.1, 1.2 and 1.3 were still
being grouped together, because they still happened on day 1, but
they weren't subsequently having their times overriden with a time
of 1)

I now apply the keyfunc to each time, and update the times of all
the fluxes in a group by the result. Since the current keyfunc is
floor() that means all the times get their decimal places chopped
off.