Fixed the first failing test in the spline integral calculation
The problem: the expected from the test was still the one from the
trapezoidal integral calculator. The spline uncertainty
calculation works differently, and so the result shouldn't be the
same as the one from the trapezoidal integration.

Specifically, with the data (in (x , y +/- dy) notation):

(0, 100 +/- 10)
(1, 100 +/- 10)

the expected integral between the data is int = 100.
If you then integrate between the upper limits of the data, you
get 110. The difference (which the code currently computes as the
uncertainty in the integral) is 110 - 100 = 10.